Continued work on DSL for Theories. Printing, constructing theorems and all kinds of definitions, authomatic conversions, and much more.
