--- src/main/scala/dev/A.scala
@@ -1,5 +1,0 @@
-package dev
-
-trait A extends WithMain {}
-
-object A extends A
