--- src/main/scala/dev/WithMain.scala
@@ -1,8 +1,0 @@
-package dev
-
-abstract class WithMain {
-  def main(args: Array[String]): Unit = {
-    println("Bonjour")
-  }
-
-}
